Move editor toolbar to separate component
